requirements.txt
requests
google-cloud-storage
apache-airflow-providers-google
google-cloud-bigquery
astronomer-cosmos